+++ playground/supabase-auth/components/ui/SubscriptionCardDemo.tsx
@@ -31,6 +31,7 @@ const SubscribeButton = () => {
 
 export const SubscriptionDemoCard = () => {
   const billing = useBilling();
+  console.log('====billing', billing);
   if (!billing.loaded) {
     return <div>Loading...</div>;
   } else if (billing.errors) {