+++ packages/nextjs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @flowglad/nextjs
 
+## 0.4.6
+
+### Patch Changes
+
+- No more find or create customer profile calls on the FlowgladContext, billing now includes a find or create
+- Updated dependencies
+  - @flowglad/server@0.4.6
+  - @flowglad/shared@0.4.6
+  - @flowglad/react@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes