+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @flowglad/react
 
+## 0.4.6
+
+### Patch Changes
+
+- No more find or create customer profile calls on the FlowgladContext, billing now includes a find or create
+- Updated dependencies
+  - @flowglad/shared@0.4.6
+
 ## 0.4.5
 
 ### Patch Changes